+++ src/pages/Home/Home.js
@@ -2,8 +2,8 @@ import React from "react";
 import { Button, Dropdown, DropdownButton } from "react-bootstrap";
 import { useNavigate } from "react-router-dom";
 
-import { FormList } from "../components/FormList";
-import Bg from "../assets/images/form-bg.png";
+import { FormList } from "../../components/FormList";
+import Bg from "../../assets/images/form-bg.png";
 
 export default function Home() {
     const navigation = useNavigate();
