+++ src/pages/Survey/Survey.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Form } from "react-bootstrap";
 import { useParams } from "react-router-dom";
-import QuestionForm from "../components/forms/QuestionForm";
+import QuestionForm from "../../components/forms/QuestionForm";
 
 import Container from "react-bootstrap/Container";
 import Button from "react-bootstrap/Button";
